site_explorer/backend/requirements.txt
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0
openai>=1.0.0
rosbags>=0.9.15
python-dotenv>=1.0.0
python-multipart>=0.0.6
Pillow>=9.0.0
opencv-python-headless>=4.7.0
numpy>=1.23.0
PyYAML>=6.0
pandas>=1.5.0
gitpython>=3.1.0
# Vector DB (optional — install separately if GPU available: faiss-gpu)
faiss-cpu>=1.7.0
sentence-transformers>=2.2.0
# Slack (optional)
slack-sdk>=3.19.0
# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0
